test: keep the FileFormatParserRegistry ordering contract in core
While this branch was open, main cleared MeshWeaver.Content.Test of its MeshWeaver.AI
reference by moving the coupled files wholesale into MeshWeaver.AI.Test. That achieves the
#2276 goal for the reference, and most of this branch became redundant on merge — reverted
here. One consequence is worth not inheriting.

ContributedParserPriorityTest travelled with the move, and it is not an AI test:
FileFormatParserRegistry is CORE. Its contributor merely happened to be the AI module's
AgentFileParser, which is what made it look like one. MeshWeaver.AI.Test is leaving this
repository with the module, so the registry's ordering contract would have left with it —
core would keep a registry whose priority behaviour nothing here verifies, and nobody would
notice, because the test would still be green somewhere else.

So the test comes back to core with a test-local ProbeFileParser. That is a strengthening,
not a concession: it is a test of the REGISTRY and should hold for every contributor,
including ones that do not exist yet. Pinning it against one module's parser had quietly
narrowed the assertion to "this one client works".

The agent-specific half stays in MeshWeaver.AI.Test as AgentParserContributionTest, where it
asserts the thing only the AI module can — that an `nodeType: Agent` file survives as
AgentConfiguration rather than degrading into a node that still claims to be an agent but
carries no instructions, delegations or plugins.

Reverted as redundant after main's move: the SpacesInFileNameTest split (both halves now sit
in the same project, so splitting them gains nothing) and the Content.Test csproj edit (main
removed the reference already).

Refs #2276
